revert refactor changes
