--- parse.go
@@ -1,141 +1,0 @@
-package cleanenv
-
-import (
-	"encoding/json"
-	"fmt"
-	"io"
-	"os"
-	"path/filepath"
-	"reflect"
-	"strings"
-
-	"github.com/BurntSushi/toml"
-	"github.com/joho/godotenv"
-	"gopkg.in/yaml.v3"
-	"olympos.io/encoding/edn"
-)
-
-// parseFile parses configuration file according to it's extension
-//
-// Currently following file extensions are supported:
-//
-// - yaml
-//
-// - json
-//
-// - toml
-//
-// - env
-//
-// - edn
-func parseFile(path string, cfg interface{}) error {
-	// open the configuration file
-	f, err := os.OpenFile(path, os.O_RDONLY|os.O_SYNC, 0)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	// parse the file depending on the file type
-	switch ext := strings.ToLower(filepath.Ext(path)); ext {
-	case ".yaml", ".yml":
-		err = parseYAML(f, cfg)
-	case ".json":
-		err = parseJSON(f, cfg)
-	case ".toml":
-		err = parseTOML(f, cfg)
-	case ".edn":
-		err = parseEDN(f, cfg)
-	case ".env":
-		err = parseENV(f, cfg)
-	default:
-		return fmt.Errorf("file format '%s' doesn't supported by the parser", ext)
-	}
-	if err != nil {
-		return fmt.Errorf("config file parsing error: %s", err.Error())
-	}
-	return nil
-}
-
-// parseYAML parses YAML from reader to data structure
-func parseYAML(r io.Reader, str interface{}) error {
-	return yaml.NewDecoder(r).Decode(str)
-}
-
-// parseJSON parses JSON from reader to data structure
-func parseJSON(r io.Reader, str interface{}) error {
-	return json.NewDecoder(r).Decode(str)
-}
-
-// parseTOML parses TOML from reader to data structure
-func parseTOML(r io.Reader, str interface{}) error {
-	_, err := toml.NewDecoder(r).Decode(str)
-	return err
-}
-
-// parseEDN parses EDN from reader to data structure
-func parseEDN(r io.Reader, str interface{}) error {
-	return edn.NewDecoder(r).Decode(str)
-}
-
-// parseENV, in fact, doesn't fill the structure with environment variable values.
-// It just parses ENV file and sets all variables to the environment.
-// Thus, the structure should be filled at the next steps.
-func parseENV(r io.Reader, _ interface{}) error {
-	vars, err := godotenv.Parse(r)
-	if err != nil {
-		return err
-	}
-
-	for env, val := range vars {
-		if err = os.Setenv(env, val); err != nil {
-			return fmt.Errorf("set environment: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// parseSlice parses value into a slice of given type
-func parseSlice(valueType reflect.Type, value string, sep string, layout *string) (*reflect.Value, error) {
-	sliceValue := reflect.MakeSlice(valueType, 0, 0)
-	if valueType.Elem().Kind() == reflect.Uint8 {
-		sliceValue = reflect.ValueOf([]byte(value))
-	} else if len(strings.TrimSpace(value)) != 0 {
-		values := strings.Split(value, sep)
-		sliceValue = reflect.MakeSlice(valueType, len(values), len(values))
-
-		for i, val := range values {
-			if err := parseValue(sliceValue.Index(i), val, sep, layout); err != nil {
-				return nil, err
-			}
-		}
-	}
-	return &sliceValue, nil
-}
-
-// parseMap parses value into a map of given type
-func parseMap(valueType reflect.Type, value string, sep string, layout *string) (*reflect.Value, error) {
-	mapValue := reflect.MakeMap(valueType)
-	if len(strings.TrimSpace(value)) != 0 {
-		pairs := strings.Split(value, sep)
-		for _, pair := range pairs {
-			kvPair := strings.SplitN(pair, ":", 2)
-			if len(kvPair) != 2 {
-				return nil, fmt.Errorf("invalid map item: %q", pair)
-			}
-			k := reflect.New(valueType.Key()).Elem()
-			err := parseValue(k, kvPair[0], sep, layout)
-			if err != nil {
-				return nil, err
-			}
-			v := reflect.New(valueType.Elem()).Elem()
-			err = parseValue(v, kvPair[1], sep, layout)
-			if err != nil {
-				return nil, err
-			}
-			mapValue.SetMapIndex(k, v)
-		}
-	}
-	return &mapValue, nil
-}
